+++ build-binary.sh
@@ -32,22 +32,9 @@ cp -f node_modules/libxmljs-mt/build/Release/*.node ${TARGET_OS}/node_modules/li
 mkdir -p ${TARGET_OS}/node_modules/libxslt/build/Release
 cp -f node_modules/libxslt/build/Release/*.node ${TARGET_OS}/node_modules/libxslt/build/Release/
 
-# Custom oc build (for linux only)
-if [ "$OS" = "linux" ] ; then
-  npm run oc-build
-  mkdir -p ${TARGET_OS}-oc
-  pkg validate --targets node${NODE_VERSION}-${OS}-x64 --output validate-${OS}-oc
-  mv ./validate-${OS}-oc ${TARGET_OS}-oc/${FILENAME}
-  cp -r ${TARGET_OS}/node_modules/libxmljs-mt/build/ ${TARGET_OS}-oc/node_modules/libxmljs-mt/build/
-  cp -r ${TARGET_OS}/node_modules/libxslt/build/ ${TARGET_OS}-oc/node_modules/libxslt/build/
-fi
-
 # Create a zip file
 cd ${TARGET_BASE}
 zip -r9 ${OS}.zip ${OS}
-if [ "$OS" = "linux" ] ; then
-  zip -r9 ${OS}-oc.zip ${OS}-oc
-fi
 
 # Test (a valid form, so it exits with 0)
 ./${OS}/${FILENAME} ../enketo-validate/test/xform/model-only.xml